Move events to API, docs, licensing
