+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/ChatBooleanEvent.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.chat.impl;
+package org.quiltmc.qsl.chat.api;
 
 import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.NotNull;
 import org.quiltmc.qsl.base.api.event.Event;
-import org.quiltmc.qsl.chat.api.QuiltMessageType;
 import org.quiltmc.qsl.chat.api.types.AbstractChatMessage;
 
 import java.util.EnumSet;
